Add ability to look up records by name.
Went with a simple regex to check if the given name occurs anywhere in the String returned by getName(). This is very broad, but probably the best solution here as names have lots of variations, and people may search by short name.

Additionally, I considered putting a minimum length requirement on this (other than > 0), but I know of two-letter names out there, and there may be one-letter names, so I concluded that was a bad idea.

Signed-off-by: Murdo B. Maclachlan <murdomaclachlan@noreply.codeberg.org>